day 3 part 2

+++ src/main/kotlin/Day03.kt
@@ -1,6 +1,7 @@
 package org.example
 
 import readInput
+import java.math.BigInteger
 
 fun main() {
     val testInput = readInput("Day03_test")
@@ -10,20 +11,20 @@ fun main() {
     println(day03_part2(testInput))
 }
 
-fun day03_part1(input: List<String>): Int {
+fun day03_part1(input: List<String>): BigInteger {
     return input.sumOf { joltage(it, 2) }
 }
 
-fun day03_part2(input: List<String>): Int {
+fun day03_part2(input: List<String>): BigInteger {
     return input.sumOf { joltage(it, 12) }
 }
 
-fun joltage(bank: String, nDigits: Int) : Int {
+fun joltage(bank: String, nDigits: Int) : BigInteger {
     var result = ""
     for(i in 0..<nDigits){
         val start = if(result.isNotEmpty()) bank.indexOf(result.last()) + 1 else 0
         val end = bank.length - nDigits + i + 1
         result += bank.substring(start, end).max()
     }
-    return result.toInt()
+    return result.toBigInteger()
 }
